+++ region.h
@@ -10,6 +10,7 @@
 #include <stdint.h>
 
 int markInvalid(void);
+int markInvalidBuf(const struct buffer *buf);
 int markInvalidSilent(void);
 
 void setMark(void);